refactor: db functions to db.js
